// src/resources/communications/infrastructure/api/handleGettingOfIncomingSms/handleGettingOfIncomingSmsRequest.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface HandleGettingOfIncomingSmsRequest {
  ToCountry: string;
  ToState: string;
  SmsMessageSid: string;
  NumMedia: string;
  ToCity: string;
  FromZip: string;
  SmsSid: string;
  FromState: string;
  SmsStatus: string;
  FromCity: string;
  Body: string;
  FromCountry: string;
  To: string;
  ToZip: string;
  NumSegments: string;
  ReferralNumMedia: string;
  MessageSid: string;
  AccountSid: string;
  From: string;
  ApiVersion: string;
}